➕ Init Post UserRegistration units

+++ angular-client/src/app/auth/store/auth.effects.ts
@@ -1,11 +1,11 @@
 import { Injectable } from "@angular/core";
 import { Router } from "@angular/router";
 import { of } from "rxjs";
-import { catchError, map, switchMap, tap, withLatestFrom } from "rxjs/operators";
+import { catchError, map, switchMap, tap } from "rxjs/operators";
 import { Actions, createEffect, ofType } from "@ngrx/effects";
 import { Store } from "@ngrx/store";
 
-import { LocalStorageService } from "src/app/_shared/services";
+import { LocalStorageService, NotificationService } from "src/app/_shared/services";
 import { AuthService } from "../services";
 import { AuthState } from "../interfaces";
 import * as AuthActions from './auth.actions';
@@ -13,53 +13,63 @@ import * as AuthActions from './auth.actions';
 @Injectable()
 export class AuthEffects {
 
+  registerUser$ = createEffect(() => this._actions$.pipe(
+    ofType(AuthActions.RegisterUser),
+    switchMap(({ dto }) => this._authService.register(dto).pipe(
+      tap(accessToken => {
+        if (accessToken) {
+          this._store.dispatch(AuthActions.GetAuthUser());
+          this._router.navigateByUrl('/auth/post-registered');
+        }
+      }),
+      map(accessToken => AuthActions.RegisterUserSuccess({ accessToken })),
+      catchError(error => of(AuthActions.RegisterUserFailure({ error })))
+    ))
+  ))
+
   login$ = createEffect(() => this._actions$.pipe(
     ofType(AuthActions.Login),
     switchMap(({ loginDto }) => this._authService.login(loginDto).pipe(
+      tap(accessToken => {
+        if (accessToken) {
+          this._store.dispatch(AuthActions.GetAuthUser());
+          this._router.navigateByUrl('/home');
+        }
+      }),
       map(accessToken => AuthActions.LoginSuccess({ accessToken })),
       catchError(error => of(AuthActions.LoginFailure({ error })))
     ))
   ))
 
-  afterLoggedIn$ = createEffect(() => this._actions$.pipe(
-    ofType(AuthActions.LoginSuccess),
-    tap(({ accessToken }) => {
-      if (accessToken) {
-        this._localStorageService.setItem('accessToken', accessToken);
-        this._store.dispatch(AuthActions.GetAuthUser());
-        this._router.navigateByUrl('/home');
-      }
-    })
-  ), { dispatch: false });
-
   getAuthUser$ = createEffect(() => this._actions$.pipe(
     ofType(AuthActions.GetAuthUser),
-    map(() => {
-      let token = this._localStorageService.getItem('accessToken') as string;
-      return token ? token : "";
-    }),
-    switchMap(accessToken => {
+    switchMap(() => {
+      const accessToken = this._authService.accessToken;
+      if (!accessToken) {
+        return of(AuthActions.GetAuthUserFailure({ error: new Error('No Token!') }))
+      }
       return this._authService.requestAuthUser().pipe(
         map((user) => AuthActions.GetAuthUserSuccess({ user, accessToken })),
         catchError(error => of(AuthActions.GetAuthUserFailure({ error })))
       )
-    }),
-    catchError(_ => {
-      return of(AuthActions.GetAuthUserFailure({ error: new Error('No Token!') }))
-    }),
+    })
   ))
 
   logout$ = createEffect(() => this._actions$.pipe(
     ofType(AuthActions.Logout),
     tap(() => {
-      this._localStorageService.removeItem('accessToken');
-      this._router.navigateByUrl('/auth/login');
+      if (this._authService.accessToken) {
+        this._localStorageService.removeItem('accessToken');
+        this._router.navigateByUrl('/auth/login');
+        this._notificationService.warn('You are logged out');
+      }
     })
   ), { dispatch: false })
 
   constructor(
     private _actions$: Actions,
     private _authService: AuthService,
+    private _notificationService: NotificationService,
     private _localStorageService: LocalStorageService,
     private _router: Router,
     private _store: Store<AuthState>,
